+++ bk/methods/every.js
