+++ bk/methods/filter.js
@@ -1,7 +1,9 @@
 const pFilter = require('p-filter')
-const { prraypromise } = require('../prraypromise')
+const prraypromise = require('../prraypromise')
 
-module.exports = function (filterer, concurrency) {
+function filter (filterer, concurrency) {
   const prom = this.then((r) => concurrency ? pFilter(r, filterer, {concurrency}) : pFilter(r, filterer))
   return prraypromise(prom)
 }
+
+module.exports = filter
