+++ bk/methods/index.js
@@ -1,0 +1,18 @@
+const map = require('./map')
+const filter = require('./filter')
+const every = require('./every')
+const some = require('./some')
+const toArray = require('./toArray')
+const reduce = require('./reduce')
+const { find, findIndex } = require('./find&findIndex')
+
+module.exports = {
+  map,
+  filter,
+  reduce,
+  every,
+  some,
+  find,
+  findIndex,
+  toArray,
+}
