+++ bk/methods/reduce.js
